Fix a fatal segment fault before Demo
The structure is same as complete version. I just find a segment fault when I run outRedirection.
I use MacOS to complete all works in time, but when I want to demo this on Windows terminal. A segment fault occurs, so I fix it and optimize some other functions to make demo process more smoothy.